--- jetty-core/jetty-deploy/src/main/java/org/eclipse/jetty/deploy/util/package-info.java
@@ -1,18 +1,0 @@
-//
-// ========================================================================
-// Copyright (c) 1995-2022 Mort Bay Consulting Pty Ltd and others.
-//
-// This program and the accompanying materials are made available under the
-// terms of the Eclipse Public License v. 2.0 which is available at
-// https://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
-// which is available at https://www.apache.org/licenses/LICENSE-2.0.
-//
-// SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
-// ========================================================================
-//
-
-/**
- * Jetty Deploy : Utilities
- */
-package org.eclipse.jetty.deploy.util;
-
