+++ jetty-ee10/jetty-ee10-quickstart/src/main/java/org/eclipse/jetty/ee10/quickstart/PreconfigureQuickStartWar.java
@@ -23,6 +23,7 @@ import org.eclipse.jetty.ee10.plus.webapp.PlusConfiguration;
 import org.eclipse.jetty.ee10.webapp.MetaInfConfiguration;
 import org.eclipse.jetty.ee10.webapp.WebAppContext;
 import org.eclipse.jetty.server.Server;
+import org.eclipse.jetty.util.URIUtil;
 import org.eclipse.jetty.util.resource.Resource;
 import org.eclipse.jetty.xml.XmlConfiguration;
 import org.slf4j.Logger;
@@ -95,7 +96,7 @@ public class PreconfigureQuickStartWar
             if (!dir.exists())
                 Files.createDirectories(dir.getPath());
 
-            URI jarUri = Resource.toJarFileUri(war.getURI());
+            URI jarUri = URIUtil.toJarFileUri(war.getURI());
             try (Resource.Mount warMount = Resource.mount(jarUri))
             {
                 // unpack contents of war to directory
