+++ jetty-ee9/jetty-ee9-webapp/src/main/java/org/eclipse/jetty/ee9/webapp/MetaInfConfiguration.java
@@ -36,6 +36,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import org.eclipse.jetty.util.FileID;
 import org.eclipse.jetty.util.IO;
 import org.eclipse.jetty.util.PatternMatcher;
 import org.eclipse.jetty.util.StringUtil;
@@ -668,7 +669,7 @@ public class MetaInfConfiguration extends AbstractConfiguration
 
         try (Stream<Path> entries = Files.walk(dir)
             .filter(Files::isRegularFile)
-            .filter(MetaInfConfiguration::isTldFile))
+            .filter(path -> FileID.isTldFile(path)))
         {
             Iterator<Path> iter = entries.iterator();
             while (iter.hasNext())
@@ -698,7 +699,7 @@ public class MetaInfConfiguration extends AbstractConfiguration
         try (Stream<Path> stream = Files.walk(mount.root().getPath()))
         {
             Iterator<Path> it = stream
-                .filter(MetaInfConfiguration::isTldFile)
+                .filter(path -> FileID.isTldFile(path))
                 .iterator();
             while (it.hasNext())
             {
@@ -707,18 +708,6 @@ public class MetaInfConfiguration extends AbstractConfiguration
             }
         }
         return tlds;
-    }
-
-    private static boolean isTldFile(Path path)
-    {
-        if (!Files.isRegularFile(path))
-            return false;
-        if (path.getNameCount() < 2)
-            return false;
-        if (!path.getName(0).toString().equalsIgnoreCase("META-INF"))
-            return false;
-        String filename = path.getFileName().toString();
-        return filename.toLowerCase(Locale.ENGLISH).endsWith(".tld");
     }
 
     protected List<Resource> findClassDirs(WebAppContext context)
@@ -871,6 +860,6 @@ public class MetaInfConfiguration extends AbstractConfiguration
 
     private boolean isFileSupported(Resource resource)
     {
-        return Resource.isArchive(resource.getURI());
+        return FileID.isArchive(resource.getURI());
     }
 }
